Move config to the top level
